Add a doltlibtool script to substitute for $(LIBTOOL)
When automake encounters a target with target-specific CFLAGS, it
generates a Makefile rule that calls $(LIBTOOL) directly rather than
using $(LTCOMPILE) or $(LTCXXCOMPILE).  To handle this case, write out
a doltlibtool script and substitute it for $(LIBTOOL).  doltlibtool
invokes doltcompile if given --mode=compile and --tag={CC,CXX}, and
falls back to libtool for any other invocation.

Thanks to Alp Toker for testing.